fix timeout for header error

+++ src/impl/unix/unix_timer.cpp
@@ -13,11 +13,11 @@ namespace impl{
 #if HAS_CLOCK_GETTIME
         struct timespec  tim;
         clock_gettime(CLOCK_REALTIME, &tim);
-        return  (uint64_t)(tim.tv_sec*1000000000LL + tim.tv_nsec);   
+        return  static_cast<uint64_t>(tim.tv_sec)*1000000000LL + static_cast<uint64_t>(tim.tv_nsec);
 #else
         struct timeval timeofday;
         gettimeofday(&timeofday,NULL);
-        return  (uint64_t)( timeofday.tv_sec*1000000000LL + timeofday.tv_usec * 1000);   
+        return  static_cast<uint64_t>( timeofday.tv_sec)*1000000000LL + static_cast<uint64_t>(timeofday.tv_usec) * 1000LL;
  #endif
     }
 }
